Add RFC 8439 test vectors
Appendix A.3 of RFC 8439 ("ChaCha20 and Poly1305 for IETF Protocols"
by Y. Nir and A. Langley) contains 11 test vectors for poly1305.
Several of the test vectors were constructed to expose common defects
in poly1305 implementations so it is good to run through them
(e.g. there is a test vector to catch the mistake of doing only a
partial reduction modulo 2^130 - 5 rather than a full reduction
immediately before adding s).  The test vectors are in a new header
file and they are exercised in example-poly1305.c.